requirements.txt
pandas==2.0.0
python-dotenv==1.0.0
pytz==2023.3
email-validator==2.0.0
